Update AreaOfObjects.cpp
Before- The objects calculations and functions were taking place in only one single class, like the calculations of Square,Rectangle,Triangle and circle were in one alone class i did it
for the sake of polymorphism because as overloading code with too many classes with non useful functions is baseless. 
After- Separate classes are not added because now the program will be able to do make much more mathematical calculations on Square,Rectangle,Triangle and circle and now Square,Rectangle,Triangle and circle have their own classes and with their own function, Did this for the sake of simplicity so that the code looks readable and easy to understand.
Will be working on code much more and new functions will be added ;)
Have a good day if you are reading this.