Update AreaOfObjects.cpp
I combined all my knowledge to make this program.
When this repository was private i gave it a lot of updates, i wrote all this code by myself as you can 
guess that because the code is very simple and straight and i kept the function and variable and classes names 
very simple and easy to understand. 

Make sure you read capabilities and limitations of the program i made. :)

WHAT THIS PROGRAM IS CAPABLE OF?
With the new code and new function added now you will be able to do more calculations such as-
Squares:
1.You can now find Area of squares.
2.You can now also find perimeter of squares.

Triangles:
1. Now you can find perimeter of triangle.
2. You can find area of triangle.
3. You can also find area of triangle using HERON's formula!

Rectangles:
1.You can find area of rectangle.
2.You can find diagonal of rectangle.
3.You can also find perimeter of rectangle.

Circle:
1.You can find area of circle.
2.You can also find circumference of circle.

LIMITATIONS ON THIS PROGRAM. Obviously with reasons.

1. This program is not cross platform, this is becase of some code or you can say some features i added like
on line 44 there is a function that is _getch(); and same usage of _getch();
on other lines also there is no cross platform replacement for _getch();
so this program will only work on windows NOT on linux or any other OS, not pog :(

2.I am still in my learning age, i used all my knowledge and tricks to make this program as good as i could
some of the problems i could not fix were
(i): there is use of tons of same code lines again. i dont know but i used a lot of code and variables again and again which was
annoying and i didnt know how can i replace it.
(ii): there is not much complexity of code and it is very very basic code, because all it does is maths.

3.Once you make a choice you cannot go back, if you chose calculations for Square you cannot go back, either
you have to exit the program (ctrl + c or ctrl + x) either put some values and get your answer. also not pog.

4.The main problem is, if user enters invalid input the code is not able to handle it and just exits on its own, yes 
there are no do or while loops you know the reason. Yeah the code will not be able to handle wrong or invalid input and 
will just crash, not pog.